extract assignment into own function
